Bugs resolvidos e área de pedidos limpa após fetch com api
