+++ gg-burguer/.vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+    "recommendations": [
+        "streetsidesoftware.code-spell-checker-portuguese-brazilian"
+    ]
+}
