+++ gg-burguer/public/index.html
@@ -16,9 +16,9 @@
     <title>React App</title>
   </head>
   <body>
-    <header class="header">
+    <!-- <header class="header">
       <img src="./images/gg-burger-header2.jpg" alt="Logo GG Burger" class="logoImg">
-      </header>
+      </header> -->
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <footer class="footer">
