+++ gg-burguer/src/pages/Kitchen.js
@@ -2,45 +2,47 @@ import React, { useState, useEffect } from "react";
 import { useHistory } from "react-router-dom";
 import { Button, Grid, Paper } from "@material-ui/core";
 import ExitToAppTwoToneIcon from '@material-ui/icons/ExitToAppTwoTone';
-import UpdateStatus from './UpdateStatus'
+import UpdateStatus from './UpdateStatus';
+import Header from '../components/Header';
 
 function Kitchen() {
-  const token = localStorage.getItem("token");
-  const idHall = localStorage.getItem("id");
-  const history = useHistory();
-  const [orders, setOrders] = useState([]);
-  const [name, setName] = useState("");
+  // const token = localStorage.getItem("token");
+  // const idHall = localStorage.getItem("id");
+  // const history = useHistory();
+  // const [orders, setOrders] = useState([]);
+  // const [name, setName] = useState("");
 
-  fetch(`https://lab-api-bq.herokuapp.com/users/${idHall}`, {
-    headers: {
-      accept: "application/json",
-      Authorization: `${token}`,
-    },
-  })
-    .then((response) => response.json())
-    .then((json) => {
-      setName(json.name);
-    });
+  // fetch(`https://lab-api-bq.herokuapp.com/users/${idHall}`, {
+  //   headers: {
+  //     accept: "application/json",
+  //     Authorization: `${token}`,
+  //   },
+  // })
+  //   .then((response) => response.json())
+  //   .then((json) => {
+  //     setName(json.name);
+  //   });
 
-  const logout = (event) => {
-    event.preventDefault();
-    localStorage.removeItem("token");
-    localStorage.removeItem("id");
-    history.push("/");
-  };
+  // const logout = (event) => {
+  //   event.preventDefault();
+  //   localStorage.removeItem("token");
+  //   localStorage.removeItem("id");
+  //   history.push("/");
+  // };
 
   return (
     <div className="kitchen-feed">
-      <Grid container spacing="2">
-          <Button
+      <Header />
+      <Grid container spacing={2}>
+          {/* <Button
             id="logoutBtn"
             size="medium"
             onClick={(event) => logout(event)}
           >
             Logout
-          </Button>
+          </Button> */}
         <Grid item xs={12}>
-          <h2 className="intro">Bem vindo(a) {name}.</h2>
+          {/* <h2 className="intro">Bem vindo(a) {name}.</h2> */}
           <h2>Preparar seguintes pedidos:</h2>
         </Grid>
             <UpdateStatus />
